separate components and views
